requirements.txt
amqp==5.1.0
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
boto3==1.21.13
botocore==1.24.13
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.21.2
fs==2.4.15
fs-s3fs==1.1.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
kombu==5.2.4
MarkupSafe==2.1.0
minio==7.1.4
model-bakery==1.4.0
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
